requirements.txt
discord.py
google-generativeai
python-dotenv
aiohttp
beautifulsoup4
lxml